% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball.R
\docType{data}
\name{basketball}
\alias{basketball}
\title{College Basketball scores}
\format{
A data frame with continously updated rows 6 variables:
\describe{
\item{date}{price, date of game}
\item{away}{the away team}
\item{away_score}{the away team score}
\item{home}{the home team}
\item{home_score}{the home team score}
\item{point_diff}{the point diffrential for the home team}
...
}
}
\source{
\url{"http://kenpom.com/cbbga22.txt"}
}
\usage{
basketball
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
