% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbb_shiny.R
\name{cbb_shiny}
\alias{cbb_shiny}
\title{Runs college basketball shiny app
must have df named basketball in environment}
\usage{
cbb_shiny()
}
\arguments{
\item{directory}{the file directory that contains ui and server}
}
\value{
college basketball shiny app
}
\description{
Runs college basketball shiny app
must have df named basketball in environment
}
\examples{
shiny_app("C:/Users/James/OneDrive/School/MATH_3190/github/hw2package/inst")
}
