% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_all.R
\name{collapse_all}
\alias{collapse_all}
\title{returns collapsed data for college basketball}
\usage{
collapse_all()
}
\value{
df with collapsed basketball data for all teams
}
\description{
returns collapsed data for college basketball
}
\examples{

collapse_all()
}
