% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_graph.R
\name{date_graph}
\alias{date_graph}
\title{Returns a graph of points scored by date}
\usage{
date_graph(team, data)
}
\arguments{
\item{team}{college basketball team to search for}

\item{data}{a df containing college baskeball games}
}
\value{
returns a ggplot graph of the points scored by selected team
}
\description{
Returns a graph of points scored by date
}
\examples{
date_graph("Southern Utah", basketball_2022)
}
