% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_graph.R
\name{game_graph}
\alias{game_graph}
\title{Returns a graph of a basketball team's point differential per game}
\usage{
game_graph(team, data)
}
\arguments{
\item{team}{college basketball team to search visualize}

\item{data}{a df containing college baskeball games}
}
\value{
a graph showing point differential by game
}
\description{
Returns a graph of a basketball team's point differential per game
}
\examples{
game_graph("Southern Utah", basketball_2022)

}
