% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{returns college basketball data}
\usage{
get_data()
}
\value{
college basketball data collected and manipulated from "http://kenpom.com/cbbga22.txt"
}
\description{
returns college basketball data
}
\examples{
get_data()
}
