\name{log.2}
\alias{log.2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Logarithms and Exponentials
}
\description{
log computes logarithms, by default natural logarithms, log10 computes common (i.e., base 10) logarithms, and log2 computes binary (i.e., base 2) logarithms. The general form log(x, base) computes logarithms with base base. log.2(x) returns the error message "negative input, NA introduced!" if negative value entered.
}
\usage{
log.2(x)
}
\arguments{
  \item{x}{

a numeric or complex vector.

base
a positive or complex number: the base with respect to which logarithms are computed. Defaults to ee=exp(1).
}
}
