% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_diff.R
\name{point_diff}
\alias{point_diff}
\title{Returns data frame of teams point diff home and away}
\usage{
point_diff(team, data)
}
\arguments{
\item{team}{college basketball team to search for}

\item{data}{a df containing college baskeball games}
}
\value{
point differential for selected team for both home and away games
}
\description{
Returns data frame of teams point diff home and away
}
\examples{
point_diff("Southern Utah", basketball_2022)
}
