\name{sqrt.2}
\alias{sqrt.2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Miscellaneous Mathematical Functions
}
\description{
abs(x) computes the absolute value of x, sqrt(x) computes the (principal) square root of x, \sqrt{x}
x
​
 .
Returns error message "negative input, NA introduced!" if negative value entered.

The naming follows the standard for computer languages such as C or Fortran.
}
\usage{
sqrt.2(x)
}
\arguments{
  \item{x}{

a numeric or complex vector or array.
}
}
