% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_select.R
\name{team_select}
\alias{team_select}
\title{Returns games played by a college basketball team}
\usage{
team_select(team, data, chr = FALSE)
}
\arguments{
\item{team}{college basketball team to search for}

\item{data}{a df containing college baskeball games}

\item{chr}{returns dates as a character if == TRUE}
}
\value{
all games played by team
}
\description{
Returns games played by a college basketball team
}
\examples{
team_select("Southern Utah", basketball_2022)
}
