% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_per.R
\name{win_per}
\alias{win_per}
\title{Returns winning percentage and win/loss record for a given college basketball team.}
\usage{
win_per(team, data)
}
\arguments{
\item{team}{college basketball team to search for}

\item{data}{a df containing college baskeball games}
}
\value{
Winning percentage as well as total wins and losses
}
\description{
Returns winning percentage and win/loss record for a given college basketball team.
}
\examples{
win_per("Southern Utah", basketball_2022)

}
