% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_operator.R
\name{f_operator}
\alias{f_operator}
\title{f_operator}
\usage{
f_operator(f)
}
\arguments{
\item{f}{hw2Package::sqrt.2(x) or hw2Package::log.2(x)}
}
\value{
the output of parameter f, or error condition object if input for f is negative
}
\description{
function operator that transform \emph{sqrt.2()} and \emph{log.2()} into functions such that:
\enumerate{
\item If input is greater than 0, then return normal output.
\item If input is invalid, return an error condition object, with "invalid_input" subclass and invalid value attached.
}
}
\examples{
log.3 = f_operator(log.2)
sqrt.3 = f_operator(sqrt.2)
x = lapply(runif(2,-10,10), log.3)
y = lapply(runif(2,-10,10), sqrt.3)
}
