% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt.2.R
\name{sqrt.2}
\alias{sqrt.2}
\title{sqrt.2}
\usage{
\method{sqrt}{`2`}(x)
}
\arguments{
\item{x}{a numeric}
}
\value{
squareroot of x or an error message if input is negative.
}
\description{
A version of sqrt() that triggers errors with a negative input
}
\examples{
sqrt.2(2)
sqrt.2(-2)
}
